+++ src/dartfx/rdf/pydantic/__init__.py
@@ -1,5 +1,5 @@
 """Public package exports for :mod:`dartfx.rdf.pydantic`."""
 
-from ._base import RdfBaseModel, RdfProperty
+from ._base import LangString, LocalizedStr, RdfBaseModel, RdfProperty
 
-__all__ = ["RdfBaseModel", "RdfProperty"]
+__all__ = ["RdfBaseModel", "RdfProperty", "LangString", "LocalizedStr"]